# backend/docs/swagger.yaml
basePath: /api
definitions:
  handlers.AdminUserResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      last_login:
        type: string
      role:
        $ref: "#/definitions/models.AdminRole"
    type: object
...
  handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
      - email
      - password
    type: object
  handlers.LoginResponse:
    properties:
      admin:
        $ref: "#/definitions/handlers.AdminUserResponse"
      message:
        type: string
      success:
        type: boolean
      token:
        type: string
    type: object
...
  models.AdminRole:
    enum:
      - super_admin
      - inventory
      - order_manager
      - viewer
    type: string
    x-enum-varnames:
      - RoleSuperAdmin
      - RoleInventory
      - RoleOrderManager
      - RoleViewer
...
host: localhost:8080
info:
  contact:
    email: admin@bjjstore.com
    name: Calvin Le
    url: http://www.bjjstore.com/support
  description: A complete e-commerce API for Brazilian Jiu-Jitsu gear and equipment
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: BJJ Store API
  version: "1.0"
paths:
  /admin/auth/login:
    post:
      consumes:
        - application/json
      description: Authenticate admin user and get JWT token
      parameters:
        - description: Admin login credentials
          in: body
          name: credentials
          required: true
          schema:
            $ref: "#/definitions/handlers.LoginRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/handlers.LoginResponse"
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Admin login
      tags:
        - admin
        - auth
  /admin/logout:
    post:
      consumes:
        - application/json
      description: Logout admin user and revoke JWT token
      produces:
        - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties: true
            type: object
        "400":
          description: No authorization header
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
        - BearerAuth: []
      summary: Admin logout
      tags:
        - admin
        - auth
...
  /admin/profile:
    get:
      consumes:
        - application/json
      description: Get current admin user profile
      produces:
        - application/json
      responses:
        "200":
          description: Admin profile
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Admin not found
          schema:
            additionalProperties: true
            type: object
      security:
        - BearerAuth: []
      summary: Get admin profile
      tags:
        - admin
        - auth
  /admin/stats:
    get:
      consumes:
        - application/json
      description: Get overview statistics for admin dashboard
      produces:
        - application/json
      responses:
        "200":
          description: Dashboard statistics
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Insufficient permissions
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
        - BearerAuth: []
      summary: Get admin dashboard statistics
      tags:
        - admin
        - dashboard
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"